requirements.txt
aiosmtplib==3.0.2
alembic==1.14.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
Cython==3.0.11
et_xmlfile==2.0.0
fastapi==0.115.4
flower==2.0.1
greenlet==3.1.1
h11==0.14.0
humanize==4.11.0
idna==3.10
kombu==5.4.2
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
vine==5.1.0
wcwidth==0.2.13
